Allow size of image to be zero when schema1 from Hub
The DockerHub now converts schema2 image manifests pushed by Docker 1.10
or higher clients into schema1, but when it does that it drops all
v1Compatibility objects except the top layer.  When it does that it
loses size info.  Since we anticipate moving to schema2 soon, and the
number of schema1 images should decline, adapt the tests to flag this
error.