Configurable deployment controller resync interval
Make the deployment controllers' resync intervals configurable. The
hard-coded 2 minute default is the likely cause of performance
regressions when thousands of deploymentconfigs are present in the
system.

When controller processing time exceeds the resync interval, the queue
will be fully replaced with a randomized list of all existing resources,
causing erratic processing times.

https://bugzilla.redhat.com/show_bug.cgi?id=1365146