router: Ensure reload on initial sync
Previously, the router wouldn't reload HAProxy after the initial sync
if the last item of the initial list of any of the watched resources
didn't reach the router to trigger a commit.  One possible trigger for
this condition was a route specifying a host already claimed by
another namespace. The router could be left in its initial state until
another commit-triggering event occurred (e.g. a watch event).

This change refactors the commit handling so that the plugin no longer
triggers a commit directly.  Instead:

 - the plugin tracks whether state has changed
 - a commit is only attempted if state has changed
 - assuming a sync is not in progress, the controller will trigger a
   commit attempt after calling one of the relevant plugin Handle methods