UPSTREAM: <carry>: Fix the IP to pod map
Change podsByIP from map[string]*kapi.Pod to map[string]kapi.Pod; in
some cases we are given a pod, not a pointer-to-pod, and while go
guarantees that "&pod" will not ever refer to uninitialized memory, it
can't guarantee that the variable won't be used to store a different
pod later. In particular, the re-use of a single local variable in
PopulatePodsByIP() meant that the map ended up getting initialized
with multiple copies of the last pod in the list, causing major
problems.