Install OpenVSwitch from the CentOS PaaS SIG Repos
In the past, we were not able to install OpenVSwitch from the normal
CentOS repositories or the `epel-release` repositories as the version
there was too old. We were using Adam Miller's COPR to source a newer
version, however this was unstable and flaky. The CentOS PaaS SIG now
provides the `openshift-origin15` repository which has the dependencies
necessary for bleeding-edge builds of Origin on CentOS, including the
correct OpenVSwitch dependency RPM. We will use this repository from
now on.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>