Add then resync then Delete cause Pop() panic
The eventqueue did not properly handle the following sequence of
operations:

handleEvent ADD default/hello
Pop         ADD default/hello
Resync
handleEvent DELETE default/hello
Pop         DELETE default/hello
Pop         MODIFIED default/hello  <<< this Pop should not happen

The last Pop references the default/hello route which has been
deleted.  This results in a panic in Pop()

The root cause of the problem is the resync adds the route to the
queue with the default ADD state instead of MODIFY.

The previous fix for this in Pop() (46a93289) was moved to Resync().

Fixes 1447928
https://bugzilla.redhat.com/show_bug.cgi?id=1447928