Don't allow deleted routes in resync list
When handleEvent() is called to ADD a route, then called again to
DELETE the route before Pop() processes the ADD, there is no need
for the route to remain in store. handleEvent() can remove it from
store. Deleted routes that are found in store during a Resync()
are deleted by Pop() as they are encouneted.

This limits the number of deleted routes in store and reduces the
number of routes that Resync() adds to the queue.