[3.5] Prevent s2i build containers from running as root user
Origin 15766a1 added logic to the SCC admission plugin to
special case the garbage collector's update of finalizers and
ownerrefs in v3.6.  This change also updated the s2i
SourceBuildStrategy.canRunAsRoot logic to pass a nil old object
to the build controller's internal admission plugin.  When this
change was backported in 2e204ff for OSE v3.5, the canRunAsRoot
change was lost.  Thus all s2i builder images run in a container
with no UID restrictions and no dropped capabilities. Note - this
container is otherwise isolated from the host, as it is not
privileged and does not mount any host data.

v3.5.5.31.30-1 and later v3.5 versions are vulnerable.

Fixes:
* Backported admission controller fixes in openshift/origin#14891
and openshift/origin#14901
* Update admission controller to handle gc case for pod updates only.

Bug 1568292