Fix the check for "pod has HostPorts"
origin PR 16692 was previously ported to OSE 900
A problem in it was fixed by origin PR 17032 which is back ported here

Origin PR16692 (OSE PR 900) didn't work because
kubehostport.PodPortMapping.PortMappings includes all ports that a pod
declares, not just hostports, so shouldSyncHostports will return true
if any pod declares that it listens on any port. (ie, always)

This (origin PR 17032) fixes that, which should make the code work in
the way origin PR16692 (OSE PR 900) intended for it to work ("only
call SyncHostPorts if there is at least one pod on the node that uses
HostPorts"). It doesn't attempt to change it to "only call SyncHostPorts
if the set of active HostPorts actually changed" because that would
be larger and less suitable as a last-minute fix.

Author:    Dan Winship <danw@redhat.com>