Change auto-egress-ip pkt_mark to be based on VNID, avoid masqueradeBit
If the iptables rules get reordered after a flush, then it's possible
that kube-proxy rules that look at MasqueradeBit will accidentally
match a packet that was intended for an auto-egress-ip rule.

To avoid that, change the code to avoid using that bit of pkt_mark.
This means we now only have 31 bits to work with instead of 32, so
make the mark be based on the (24-bit) VNID rather than the (32-bit)
egress IP address.