Do not retry conflicts during storage migration
Storage migration is special in that all it needs to do is a no-op
update to cause the API server to migrate the object to the
preferred version.  Thus if we encounter a conflict, we know that
something updated the object and we no longer need to do anything.
If the object needed migration, the API server has already migrated
it.

Signed-off-by: Monis Khan <mkhan@redhat.com>